feat: add new pages `service`, `gelangensnachweis`
Every page must to be listed in `src/params/page.ts`. This was a
deliberate design decision back then: Pages are hard-coded in the source
while blog posts can be added arbitrarily by Rhenania's staff.

It doesn't *have* to be this way, we could also make a request on every
`/url/<slug>` and only render a 404 if there is no page. We do the same
for blog posts.