feat: implement server-side search
 ## Changes

This one took me much more time than expected. The implementation was
super smooth and quickly done (< 30min) but the cursed automatic update
of the search index took hours. It's simply impossible to run just a
single file of a svelte project with e.g. `ts-node`. Multiple
configurations are conflicting with each other.

In the end I found no better way than the maintainer of `svelte-algolia`
and using `src/hooks.ts` with an if-clause:
https://github.com/janosh/svelte-algolia