fix: replaceAllObjects
This was a classic "Read the f** manual": There was already a dedicated
API request in Algolia which does exactly what we want.

For us this means the procedure is **atomic** meaning, the index is
either successfully replaced or nothing happened. There is no more
in-between state, ie. the index is deleted.