fix: don't show `false` on initial page load
This fixes a minor bug in production where you would see `false` in the
search bar for a few seconds.

I didn't know that "null coalescing operator" `??` only coaleses `null` and
`undefined`, not `false`. So, during prerendering, the left hand side
will be `false`, thus the whole expression is `false`. In the template it will
be rendered as string `"false"`.