feat: render JSON-LD for every AccordionContainer
Ok, this will render a JSON-LD for *every* AccordionContainer similar to this example:
https://developers.google.com/search/docs/appearance/structured-data/faqpage#examples

Note that wether or not the `AccordionContainer` contains FAQ it will render them as JSON-LD. Maybe it would be better to have a dedicated component for FAQs?

This PR will also introduce a bug: Every JSON-LD will be rendered twice. This is most likely a bug in the current version of svelte-kit:
https://github.com/sveltejs/svelte/pull/7745

Updating the dependency should fix the problem.